requirements.txt
gunicorn==19.9.0
asn1crypto==0.24.0
Authlib==0.11
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dominate==2.4.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-OAuth==0.12
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
oauth2==1.9.0.post1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
python-dotenv==0.10.3
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.8
uritemplate==3.0.0
urllib3==1.25.4
virtualenv==16.7.5
visitor==0.1.3
Werkzeug==0.15.6
WTForms==2.2.1
